edge-matrix/miner/helper/crypto.go
package helper

import (
	"github.com/emc-protocol/edge-matrix/crypto"
	"github.com/emc-protocol/edge-matrix/types"
	"math/big"
)

// Magic numbers from Ethereum, used in v calculation
var (
	big27 = big.NewInt(27)
	big35 = big.NewInt(35)
)

func CalculateV(parity byte, chainID uint64) []byte {
	reference := big.NewInt(int64(parity))
	reference.Add(reference, big35)

	mulOperand := big.NewInt(0).Mul(big.NewInt(int64(chainID)), big.NewInt(2))

	reference.Add(reference, mulOperand)

	return reference.Bytes()
}

// ecrecover recovers signer address from the given digest and signature
func ecrecover(sig, msg []byte) (types.Address, error) {
	pub, err := crypto.RecoverPubkey(sig, msg)
	if err != nil {
		return types.Address{}, err
	}

	return crypto.PubKeyToAddress(pub), nil
}
